update test into demos
