+++ src/main/java/net/clesperanto/demos/YetAnotherPlayground.java
@@ -30,34 +30,43 @@
  * OF THE POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package net.clesperanto.test;
+package net.clesperanto.demos;
+
+import ij.IJ;
+import ij.ImageJ;
+import ij.ImagePlus;
 
 import net.clesperanto.CLIJ3;
 import net.clesperanto.core.ArrayJ;
 
-import net.imagej.Dataset;
-import net.imagej.ImageJ;
-import net.imglib2.RandomAccessibleInterval;
-import net.imglib2.img.Img;
-import net.imglib2.type.numeric.real.FloatType;
+public class YetAnotherPlayground {
+    public static void main(String[] args) {
 
-public class YetAnotherPlaygroundImgLib2 {
+        new ImageJ();
 
-    final static net.imagej.ImageJ ij = new ImageJ();
-
-    public static void main(String[] args) throws Exception {
-        ij.launch(args);
-
-        Dataset dataset = (Dataset) ij.io().open("./demo/boats.tif");
-
-        Img<FloatType> img = ij.op().convert().float32((Img) dataset);
-        ij.ui().show("input", img);
+        ImagePlus imp = IJ.openImage("./demo/blobs.tif");
+        imp.show();
 
         CLIJ3 cle = CLIJ3.getInstance();
-        ArrayJ output = cle.gaussian_blur(img, null, 3, 0, 0);
-        cle.imshow(output);
 
-        RandomAccessibleInterval result = cle.pullRAI(output);
-        ij.ui().show("output", result);
+        ArrayJ blurred = cle.gaussian_blur(imp, null, 10, 0, 0);
+        ArrayJ binary = cle.threshold_otsu(blurred, null);
+        ArrayJ labels = cle.connected_components_labeling(binary, null, "box");
+
+        cle.imshow(labels);
+        ImagePlus result = cle.pull(labels);
+
+        // now print information about the result in ImageJ console
+        IJ.log("ImagePlus object: " + result);
+        IJ.log("Image dimensions: " + result.getWidth() + "x" + result.getHeight() + "x" + result.getNSlices());
+        IJ.log("Image type: " + result.getType());
+        IJ.log("Image title: " + result.getTitle());
+
+
+        // now print information about the result in ImageJ console
+        IJ.log("ArrayJ object: " + labels);
+        IJ.log("ArrayJ dimensions: " + labels.width() + "x" + labels.height() + "x" + labels.depth());
+        IJ.log("ArrayJ type: " + labels.dataType());
+        
     }
 }
