+++ src/main/java/net/clesperanto/demos/YetAnotherPlaygroundImgLib2.java
@@ -30,43 +30,34 @@
  * OF THE POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package net.clesperanto.test;
-
-import ij.IJ;
-import ij.ImageJ;
-import ij.ImagePlus;
+package net.clesperanto.demos;
 
 import net.clesperanto.CLIJ3;
 import net.clesperanto.core.ArrayJ;
 
-public class YetAnotherPlayground {
-    public static void main(String[] args) {
+import net.imagej.Dataset;
+import net.imagej.ImageJ;
+import net.imglib2.RandomAccessibleInterval;
+import net.imglib2.img.Img;
+import net.imglib2.type.numeric.real.FloatType;
 
-        new ImageJ();
+public class YetAnotherPlaygroundImgLib2 {
 
-        ImagePlus imp = IJ.openImage("./demo/blobs.tif");
-        imp.show();
+     final static net.imagej.ImageJ ij = new ImageJ();
+
+    public static void main(String[] args) throws Exception {
+        ij.launch(args);
+
+        Dataset dataset = (Dataset) ij.io().open("./demo/boats.tif");
+
+        Img<FloatType> img = ij.op().convert().float32((Img) dataset);
+        ij.ui().show("input", img);
 
         CLIJ3 cle = CLIJ3.getInstance();
+        ArrayJ output = cle.gaussian_blur(img, null, 3, 0, 0);
+        cle.imshow(output);
 
-        ArrayJ blurred = cle.gaussian_blur(imp, null, 10, 0, 0);
-        ArrayJ binary = cle.threshold_otsu(blurred, null);
-        ArrayJ labels = cle.connected_components_labeling(binary, null, "box");
-
-        cle.imshow(labels);
-        ImagePlus result = cle.pull(labels);
-
-        // now print information about the result in ImageJ console
-        IJ.log("ImagePlus object: " + result);
-        IJ.log("Image dimensions: " + result.getWidth() + "x" + result.getHeight() + "x" + result.getNSlices());
-        IJ.log("Image type: " + result.getType());
-        IJ.log("Image title: " + result.getTitle());
-
-
-        // now print information about the result in ImageJ console
-        IJ.log("ArrayJ object: " + labels);
-        IJ.log("ArrayJ dimensions: " + labels.width() + "x" + labels.height() + "x" + labels.depth());
-        IJ.log("ArrayJ type: " + labels.dataType());
-        
+        RandomAccessibleInterval result = cle.pullRAI(output);
+        ij.ui().show("output", result);
     }
 }
